modules/person_api/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements/production.in
#
aniso8601==9.0.1
    # via flask-restx
attrs==21.2.0
    # via jsonschema
click==8.0.1
    # via flask
colorama==0.4.4
    # via click
flask==2.0.1
    # via
    #   -r requirements\common.in
    #   flask-cors
    #   flask-restx
    #   flask-sqlalchemy
flask-accepts==0.18.2
    # via -r requirements\common.in
flask-cors==3.0.10
    # via -r requirements\common.in
flask-restx==0.5.1
    # via
    #   -r requirements\common.in
    #   flask-accepts
flask-sqlalchemy==2.5.1
    # via -r requirements\common.in
greenlet==1.1.1
    # via sqlalchemy
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
jsonschema==3.2.0
    # via flask-restx
markupsafe==2.0.1
    # via jinja2
marshmallow==3.13.0
    # via
    #   flask-accepts
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.26.1
    # via -r requirements\common.in
psycopg2-binary==2.9.1
    # via -r requirements\common.in
pyrsistent==0.18.0
    # via jsonschema
python-dotenv==0.19.0
    # via -r requirements\common.in
pytz==2021.1
    # via flask-restx
six==1.16.0
    # via
    #   flask-cors
    #   flask-restx
    #   jsonschema
sqlalchemy==1.4.25
    # via
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
waitress==2.0.0
    # via -r requirements/production.in
werkzeug==2.0.1
    # via
    #   flask
    #   flask-accepts
    #   flask-restx

# The following packages are considered to be unsafe in a requirements file:
# setuptools
